Got verification tests working.  Note: very kludgy solution for verification tests, but going with it for now until I figure out a better solution

+++ tree_intersection/C++/tests/verification/repeat_values/test_repeat_values.cpp
@@ -1,8 +1,26 @@
 #include "gtest/gtest.h"
 #include <stdlib.h>
+#include <cstdio>
+#include <iostream>
+#include <memory>
+#include <stdexcept>
+#include <string>
+#include <array>
 
 TEST(TestRepeatValues, test6) {
+
+    std::system("../source/tree_intersection_run ../../tests/verification/repeat_values/input.json");
+
+    std::array<char, 128> buffer;
+    std::string differences;
+    std::unique_ptr<FILE, decltype(&pclose)> pipe(popen("diff ./output.json ../../tests/verification/repeat_values/gold/output.json", "r"), pclose);
+    if (!pipe) {
+        throw std::runtime_error("popen() failed!");
+    }
+    while (fgets(buffer.data(), buffer.size(), pipe.get()) != nullptr) {
+        differences += buffer.data();
+    }
+
+    EXPECT_TRUE(differences.empty());
     std::system("rm ./output.json");
-    std::system("../../../build/source/tree_intersection input.json");
-    EXPECT_TRUE(std::system("diff ./ouput.json ./gold/output.json"));
 }
